BL31: Drop frequency initialization
Moving the initialization of the memory mapped frequency register
CNTFID to BL2 in the patch "BL2: reset: Enable generic counter" we
can drop that from the BL31 code.

To ensure that the initialization moved to BL2 does the correct thing,
add some checks to BL31. These are active in DEBUG build only, so don't
affect the release product build:

- Check if the memory mapped frequency is calculated correctly
- cntfrq_el0 must match the frequency read from memory mapped space
- Neither cntfrq_el0, nor memory mapped frequency can be zero

Signed-off-by: Dirk Behme <dirk.behme@de.bosch.com>