plat: rcar: io storage: drv_policies is rw
drv_policies is marked as const and with this placed by the linker
in the .ro section. This is wrong, as in io_rcar.c in the function
load_bl33x() there is a write access to this struct:

/* set eMMC partition */
((io_drv_spec_t *)backend_image_spec)->partition =
	(uint32_t)emmc_prttn;	/* needs 32-bits only */

Due to the casting the linker can't know about this, so explicitly
put this struct into the .data section.

In case the MMU is enabled this fixes a data write abort to ro data.

Signed-off-by: Dirk Behme <dirk.behme@de.bosch.com>