plat: rcar: BL2: Correct MMU configuration
Correct the MMU configuration for BL2: Using DEVICE_RCAR_BASE2 - BL2_BASE
for size calculation does work only if BL2_BASE is located in internal
SYSRAM of the device. I.e. if we want to map everyting from SYSRAM start
address (0xE6300000) up to the start of the second device base (0xE6360000).
If we place the BL2 anywhere else, i.e. not in the SYSRAM, but e.g. into main
memory (e.g. at 0x43F40000) this calculation doesn't work any more. Instead,
we have to use the end of BL2 as given by the linker.

Signed-off-by: Dirk Behme <dirk.behme@de.bosch.com>