rcar-gen4: plat: BL31: add call rcar_mssr_setup in bl31_platform_setup
In R-Car V4M, INTC-AP to control the GIC is in reset and clocked off
when power is turned on. Therefore, the GIC registers cannot be accessed
and halt.

This patch solves this issue by the call rcar_mssr_setup at the top of
bl31_platform_setup, which releases the reset and turns on the clock for
required modules.

Signed-off-by: Takeshi Kihara <takeshi.kihara.df@renesas.com>