rcar_gen3: Fix issue unable to access system RAM after warm boot
Issue:
After a warm boot (Suspend-to-RAM), system RAM in the address
range 0xE6300400–0xE6360000 becomes inaccessible in Linux.
This can be observed in the kernel log as access errors when reading
or writing to this range by devmem2

Cause:
The SPTCRn registers, which configure the security protection settings
for this RAM region, are not reconfigured after a warm boot. As a
result, access remains restricted and Linux is unable to use the
protected area.

Fix:
Call bl2_ram_security_setting_finish() from bl2_secure_setting()
to reconfigure SPTCRn during resume. This restores access permissions
to the protected system RAM region after Suspend-to-RAM."

Signed-off-by: Khuong Nguyen <khuong.nguyen.rb@renesas.com>